Fixed issues with reactour not build due to dependency issues. Refactored and move Label into atoms
