+++ package.json
@@ -58,12 +58,12 @@
     "lodash": "4.17.15",
     "moment-timezone": "0.5.13",
     "prop-types": "15.6.2",
-    "react": "16.3.0",
+    "react": "16.3.2",
     "react-aria-offcanvas": "^1.3.6",
     "react-autosuggest": "9.4.2",
     "react-burger-menu": "2.6.11",
     "react-csv": "1.0.18",
-    "react-dom": "16.3.0",
+    "react-dom": "16.3.2",
     "react-i18next": "7.0.0",
     "react-image": "1.0.1",
     "react-infinite-scroller": "1.1.1",
@@ -85,7 +85,7 @@
     "react-toastify": "2.2.0",
     "react-tooltip": "3.3.1",
     "react-transition-group": "1.2.1",
-    "reactour": "1.15.4",
+    "reactour": "1.7.1",
     "redux": "^3.7.1",
     "redux-thunk": "2.3.0",
     "reselect": "^3.0.1",
