+++ src/views/atoms/icon/icon.spec.js
